requirements.txt
numpy==1.21.5
opencv_python==4.5.1.48
Pillow==9.4.0
torch==1.7.1+cu101
torchvision==0.8.2+cu101
tqdm==4.42.1
